moving some bits around
