+++ dd-java-agent/agent-debugger/src/main/java/com/datadog/debugger/probe/DebugSessionStatus.java
@@ -1,0 +1,11 @@
+package com.datadog.debugger.probe;
+
+public enum DebugSessionStatus {
+  NONE,
+  ACTIVE,
+  DISABLED;
+
+  public boolean isDisabled() {
+    return this == DISABLED;
+  }
+}
